+++ BetterZeeLog/CHANGELOG.md
@@ -4,7 +4,7 @@
 
   * Updated for the `v0.220.3` patch.
   * Added a `PlatformUserID.Constructor()` transpiler patch to filter out all the failed parsing log spam.
-  * Copied the `Container.RPC_RequestOpen()` transpiler patches for `RPC_RequestStack()` and `RPC_RequestTakeAll()`._
+  * Copied the `Container.RPC_RequestOpen()` transpiler patches for `RPC_RequestStack()` and `RPC_RequestTakeAll()`.
 
 ### 1.11.0